+++ packages/board-server/src/app/elements/elements.ts
@@ -1,0 +1,8 @@
+/**
+ * @license
+ * Copyright 2024 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { ActivityLogLite } from "./activity-log/activity-log-lite.js";
+export { AppNav } from "./nav/nav.js";