+++ packages/breadboard/src/inspector/run/run.ts
@@ -23,7 +23,7 @@ import {
   InspectableRunNodeEvent,
   InspectableRunInputs,
 } from "../types.js";
-import { DataStore } from "../../data/types.js";
+import { DataStore, RunTimestamp, RunURL } from "../../data/types.js";
 
 const isInput = (
   event: InspectableRunEvent
@@ -40,101 +40,108 @@ export class RunObserver implements InspectableRunObserver {
   #options: RunObserverOptions;
   #runs: Run[] = [];
   #runLimit = 2;
-  #restoreRuns: Promise<void> = Promise.resolve();
+  #url: RunURL | null = null;
+  #timestamp: RunTimestamp | null = null;
 
   constructor(store: GraphDescriptorStore, options: RunObserverOptions) {
     this.#store = store;
     this.#options = options;
-
-    if (this.#options.runStore) {
-      this.#restoreRuns = this.#options.runStore
-        .getNewestRuns(this.#runLimit)
-        .then((runs) => {
-          for (let i = 0; i < runs.length; i++) {
-            const results = runs[i];
-
-            let run!: Run;
-            for (const result of results) {
-              if (result.type === "graphstart") {
-                const { path, timestamp } = result.data;
-                if (path.length === 0) {
-                  run = new Run(
-                    timestamp,
-                    this.#store,
-                    result.data.graph,
-                    this.#options
-                  );
-
-                  if (!this.#options.skipDataStore) {
-                    this.#options.dataStore?.createGroup(run.dataStoreKey);
-                  }
-                }
-              } else if (result.type === "graphend") {
-                const { path, timestamp } = result.data;
-                if (path.length === 0) {
-                  run.end = timestamp;
-                }
-              }
-
-              if (!run) {
-                console.warn("Unable to restore run");
-              }
-
-              if (!this.#options.skipDataStore) {
-                this.#options.dataStore?.replaceDataParts(
-                  run.dataStoreKey,
-                  result
-                );
-              }
-              run.addResult(result);
-            }
-
-            this.#runs.push(run);
-          }
-        });
-    }
   }
 
   async runs(): Promise<InspectableRun[]> {
     return this.#runs;
   }
 
-  async #storeInRunStore(dataKey: string, result: HarnessRunResult) {
+  async #convertRunInfoToRuns(
+    url: RunURL,
+    runInfo: Map<RunTimestamp, HarnessRunResult[]>
+  ): Promise<Run[]> {
+    const runs = await Promise.all(
+      [...runInfo]
+        .filter(([, events]) => events.length > 0)
+        .sort(([timeA], [timeB]) => {
+          return timeB - timeA;
+        })
+        .map(async ([timestamp, results]) => {
+          let run!: Run;
+
+          for (const result of results) {
+            if (result.type === "graphstart") {
+              const { path } = result.data;
+              if (path.length === 0) {
+                run = new Run(
+                  timestamp,
+                  this.#store,
+                  result.data.graph,
+                  this.#options
+                );
+
+                run.dataStoreKey = `${url}-${timestamp}`;
+                if (!this.#options.skipDataStore) {
+                  // Ensure that we release old blobs if we've encountered this
+                  // run before.
+                  this.#options.dataStore?.releaseGroup(run.dataStoreKey);
+                  this.#options.dataStore?.createGroup(run.dataStoreKey);
+                }
+              }
+            } else if (result.type === "graphend") {
+              const { path, timestamp } = result.data;
+              if (path.length === 0) {
+                run.end = timestamp;
+              }
+            }
+
+            if (!run) {
+              console.warn("Unable to restore run");
+            }
+
+            if (!this.#options.skipDataStore) {
+              await this.#options.dataStore?.replaceDataParts(
+                run.dataStoreKey,
+                result
+              );
+            }
+
+            run.addResult(result);
+          }
+
+          return run;
+        })
+    );
+
+    return runs;
+  }
+
+  async #storeInRunStore(
+    url: RunURL,
+    timestamp: RunTimestamp,
+    result: HarnessRunResult
+  ) {
     const { runStore } = this.#options;
-    if (!runStore) {
+    if (!runStore || !this.#url || !this.#timestamp) {
       return Promise.resolve();
     }
 
-    if (result.type === "graphstart") {
-      const { path } = result.data;
-      if (path.length === 0) {
-        const dataKey = Date.now().toFixed(0);
-        await runStore?.start(dataKey, this.#runLimit);
-      }
-
-      await runStore?.write(result);
-    } else if (result.type === "graphend") {
-      await runStore?.write(result);
-
-      const { path, timestamp } = result.data;
-      const run = this.#runs[0];
-      if (path.length === 0) {
-        await runStore?.stop();
-        run.end = timestamp;
-      }
-    } else if (result.type === "error") {
-      await runStore?.abort();
-    } else if (result.type !== "end") {
-      await runStore?.write(result);
+    if (result.type === "error") {
+      await runStore?.abort(url, timestamp);
+    } else {
+      await runStore?.write(url, timestamp, result);
     }
   }
 
   async observe(result: HarnessRunResult): Promise<void> {
-    await this.#restoreRuns;
-
     if (result.type === "graphstart") {
       const { path, timestamp } = result.data;
       if (path.length === 0) {
+        this.#url = result.data.graph.url ?? "no-url-graph";
+        if (this.#options.runStore) {
+          this.#timestamp = await this.#options.runStore.start(this.#url);
+          const runInfo = await this.#options.runStore.getStoredRuns(this.#url);
+          this.#runs = await this.#convertRunInfoToRuns(this.#url, runInfo);
+        } else {
+          this.#timestamp = timestamp;
+        }
+
         const run = new Run(
           timestamp,
           this.#store,
@@ -148,34 +155,36 @@ export class RunObserver implements InspectableRunObserver {
 
         this.#runs.unshift(run);
 
-        if (this.#runs.length > this.#runLimit) {
-          const groupIds = this.#runs
-            .slice(this.#runLimit)
-            .map((run) => run.dataStoreKey);
-          for (const groupId of groupIds) {
-            if (this.#options.skipDataStore) {
-              continue;
-            }
-
-            this.#options.dataStore?.releaseGroup(groupId);
-          }
-          this.#runs.length = this.#runLimit;
+        if (this.#options.runStore) {
+          await this.#options.runStore.truncate(this.#url, this.#runLimit);
         }
       }
     }
 
     const run = this.#runs[0];
     if (!run) {
-      console.warn("No run available");
+      console.warn(`No run available to store ${result.type} - stopping`);
       return;
     }
 
+    if (!this.#url || !this.#timestamp) {
+      console.warn("No URL or timestamp set for the current run - stopping");
+      return;
+    }
+
+    if (result.type === "graphend") {
+      const { path, timestamp } = result.data;
+      if (path.length === 0) {
+        run.end = timestamp;
+      }
+    }
+
     if (!this.#options.skipDataStore) {
-      this.#options.dataStore?.replaceDataParts(run.dataStoreKey, result);
+      await this.#options.dataStore?.replaceDataParts(run.dataStoreKey, result);
     }
 
     run.addResult(result);
-    await this.#storeInRunStore(run.dataStoreKey, result);
+    await this.#storeInRunStore(this.#url, this.#timestamp, result);
   }
 
   async load(
@@ -193,7 +202,7 @@ export class RunObserver implements InspectableRunObserver {
 }
 
 export class Run implements InspectableRun {
-  public readonly dataStoreKey = crypto.randomUUID();
+  public dataStoreKey: string = crypto.randomUUID();
 
   #events: EventManager;