+++ packages/breadboard/src/run/run-graph.ts
@@ -32,7 +32,7 @@ export async function* runGraph(
   resumeFrom?: TraversalResult
 ): AsyncGenerator<BreadboardRunResult> {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { inputs, ...context } = args;
+  const { inputs, start, ...context } = args;
   const { probe, state, invocationPath = [] } = context;
 
   const lifecycle = state?.lifecycle();
@@ -40,6 +40,8 @@ export async function* runGraph(
     const nodeInvoker = new NodeInvoker(args, graph, next);
 
     lifecycle?.dispatchGraphStart(graph.url!, invocationPath);
+
+    let invocationId = 0;
 
     const reanimation = state?.reanimation();
     if (reanimation) {
@@ -50,14 +52,18 @@ export async function* runGraph(
           // This can only happen when `runGraph` is called by `invokeGraph`,
           // which means that all we need to do is provide the output and
           // return.
-          const { result, invocationId, path } = frame.replay();
-          await next(new OutputStageResult(result, invocationId, path));
+          const { result, invocationId: id, path } = frame.replay();
+          await next(new OutputStageResult(result, id, path));
           // The nodeend and graphend will be dispatched by `invokeGraph`.
           return;
         }
         case "resume": {
           const { result, invocationPath } = frame.resume();
+
           resumeFrom = result;
+          // Adjust invocationId to match the point from which we are resuming.
+          invocationId = invocationPath[invocationPath.length - 1];
+
           await lifecycle?.dispatchNodeStart(result, invocationPath);
 
           let outputs: OutputValues | undefined = undefined;
@@ -67,21 +73,35 @@ export async function* runGraph(
             outputs = await nodeInvoker.invokeNode(result, invocationPath);
             result.outputs = outputs;
             resumeFrom = result;
+          } else {
+            outputs = result.outputs;
           }
 
           lifecycle?.dispatchNodeEnd(outputs, invocationPath);
+
+          await probe?.report?.({
+            type: "nodeend",
+            data: {
+              node: result.descriptor,
+              inputs: result.inputs,
+              outputs: outputs as OutputValues,
+              path: invocationPath,
+              timestamp: timestamp(),
+            },
+          });
         }
       }
     }
 
-    let invocationId = 0;
     const path = () => [...invocationPath, invocationId];
 
-    const machine = new TraversalMachine(graph, resumeFrom);
-    await probe?.report?.({
-      type: "graphstart",
-      data: { graph, path: invocationPath, timestamp: timestamp() },
-    });
+    const machine = new TraversalMachine(graph, resumeFrom, start);
+    if (!resumeFrom) {
+      await probe?.report?.({
+        type: "graphstart",
+        data: { graph, path: invocationPath, timestamp: timestamp() },
+      });
+    }
 
     for await (const result of machine) {
       context?.signal?.throwIfAborted();
@@ -102,6 +122,18 @@ export async function* runGraph(
           },
         });
         continue;
+      } else {
+        lifecycle?.dispatchEdge(result.current);
+        await probe?.report?.({
+          type: "edge",
+          data: {
+            edge: result.current,
+            to: path(),
+            from: lifecycle?.pathFor(result.current.from),
+            timestamp: timestamp(),
+            value: inputs,
+          },
+        });
       }
 
       await lifecycle?.dispatchNodeStart(result, path());